Fix for construction of GDS headers in case of Mercator, Polarstereo, and latlon. Changed lTemp from long to int. Had been done already for Lambert Conf. JM
git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1528 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ external/io_grib1/MEL_grib1/gribputgds.c
@@ -1082,7 +1082,7 @@ int   inp2grib_PolarSt  (ppvGDS_Proj_Input, Polar, lProj_size , errmsg)
 * F.1       INIT variables !default stat=good
 */
    GDS_PS_INPUT      *pProjInp = 0;
-   long              lTemp = 0;
+   int               lTemp = 0;
    int               nStatus = 0;
    char		     *func="inp2grib_PolarSt";
 
@@ -1257,7 +1257,7 @@ int    inp2grib_Latlon  (ppvGDS_Proj_Input, pLatlon, lProj_size, errmsg)
 #endif
 {
    GDS_LATLON_INPUT  *Inp = 0;
-   long              lTemp = 0;
+   int               lTemp = 0;
    char 	*func= "inp2grib_Latlon";
 /*
 *
@@ -1537,7 +1537,7 @@ int   inp2grib_Mercator (ppvGDS_Proj_Input, Mercator, lProj_size , errmsg)
 * F.1       INIT variables !default stat=good
 */
    mercator          *ProjInp = 0;
-   long              lTemp = 0;
+   int               lTemp = 0;
    int               nStatus = 0;
    char		     *func="inp2grib_PolarSt";
